+++ kousa/test/beef/user_test.exs
@@ -101,7 +101,7 @@ defmodule Kousa.Beef.UserTest do
       Follows.insert(%{userId: user1.id, followerId: user2.id})
       Factory.create(Room, creatorId: user1.id)
       room = Factory.create(Room, creatorId: user2.id)
-      Kousa.Data.RoomBlock.insert(%{roomId: room.id, userId: user1.id, modId: user2.id})
+      Beef.RoomBlocks.insert(%{roomId: room.id, userId: user1.id, modId: user2.id})
       UserBlocks.insert(%{userIdBlocked: user1.id, userId: user2.id})
       Beef.RoomPermissions.ask_to_speak(user1.id, room.id)
       assert {:ok, _} = Users.delete(user1.id)
